+++ test/lib/heuristics/scheduling_test.rb
@@ -553,6 +553,20 @@ class HeuristicTest < Minitest::Test
       refute_includes result[:routes].find{ |r| r[:activities].any?{ |stop| stop[:service_id] == 'service_6_1_1' } }[:vehicle_id], 'vehicle_0_'
     end
 
+    def test_skills_in_scheduling
+      vrp = VRP.lat_lon_scheduling_two_vehicles
+      result = OptimizerWrapper.wrapper_vrp('ortools', { services: { vrp: [:ortools] }}, TestHelper.create(vrp), nil)
+      assigned_route = result[:routes].find{ |r| r[:activities].any?{ |stop| stop[:service_id] == 'service_6_1_1' } }
+      assert_includes assigned_route[:vehicle_id], 'vehicle_0_' # default result
+
+      vrp = VRP.lat_lon_scheduling_two_vehicles
+      vrp[:vehicles][1][:skills] = [[:compatible]]
+      vrp[:services].find{ |s| s[:id] == 'service_6' }[:skills] = [:compatible]
+      result = OptimizerWrapper.wrapper_vrp('ortools', { services: { vrp: [:ortools] }}, TestHelper.create(vrp), nil)
+      assigned_route = result[:routes].find{ |r| r[:activities].any?{ |stop| stop[:service_id] == 'service_6_1_1' } }
+      refute_includes assigned_route[:vehicle_id], 'vehicle_0_'
+    end
+
     def test_with_activities
       vrp = VRP.lat_lon_scheduling_two_vehicles
       vrp[:configuration][:resolution][:minimize_days_worked] = true